+++ app.ts
@@ -15,7 +15,7 @@ import { electionResultsRouter } from './routes/verkiezingsresultaten';
 import { rangordeRouter } from './routes/rangorde';
 
 import { cronjob as notificationActiveMandateesWithoutBesluitCron } from './cron/notification-for-bekrachtigde-mandataris';
-import { cronjob as processHarvestedDecisions } from './cron/handle-decision-queue';
+import { cronjob as processHarvestedDecisions } from './cron/auto-bekrachtig-mandatarissen';
 
 app.use(
   bodyParser.json({
