+++ data-access/mandataris.ts
@@ -514,90 +514,6 @@ export async function endExistingMandataris(
   }
 }
 
-export async function findDecisionAndLinkForMandataris(
-  mandatarisUri: string,
-): Promise<{ besluit: string | undefined; link: string | undefined }> {
-  const mandatarisSubject = sparqlEscapeUri(mandatarisUri);
-  const besluiteQuery = `
-  PREFIX mandaat: <http://data.vlaanderen.be/ns/mandaat#>
-  PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
-
-   SELECT ?artikel ?link
-   WHERE {
-      GRAPH ?g {
-        VALUES ?link {
-          mandaat:bekrachtigtAanstellingVan
-          mandaat:bekrachtigtOntslagVan
-        }
-        ?artikel ?link ${mandatarisSubject}.
-      }
-      OPTIONAL {
-        ?g ext:ownedBy ?eenheid.
-      }
-      FILTER(BOUND(?eenheid) || ?g = ${sparqlEscapeUri(BESLUIT_STAGING_GRAPH)})
-    }
-  `;
-
-  const result = await querySudo(besluiteQuery);
-  const sparqlResult = findFirstSparqlResult(result);
-
-  return {
-    besluit: sparqlResult?.artikel?.value,
-    link: sparqlResult?.link?.value,
-  };
-}
-
-export async function updatePublicationStatusOfMandataris(
-  mandataris: string,
-  status: PUBLICATION_STATUS,
-): Promise<void> {
-  const escaped = {
-    mandataris: sparqlEscapeUri(mandataris),
-    status: sparqlEscapeUri(status),
-    mandatarisType: sparqlEscapeTermValue(TERM_MANDATARIS_TYPE),
-  };
-  const updateStatusQuery = `
-    PREFIX lmb: <http://lblod.data.gift/vocabularies/lmb/>
-    PREFIX dct: <http://purl.org/dc/terms/>
-
-    DELETE {
-      GRAPH ?graph {
-        ${escaped.mandataris} lmb:hasPublicationStatus ?status.
-        ${escaped.mandataris} dct:modified ?modified.
-      }
-    }
-    INSERT {
-      GRAPH ?graph {
-        ${escaped.mandataris} lmb:hasPublicationStatus ${escaped.status}.
-        ${escaped.mandataris} dct:modified ?now.
-      }
-    }
-    WHERE {
-      GRAPH ?graph {
-        ${escaped.mandataris} a ${escaped.mandatarisType}.
-        OPTIONAL {
-          ${escaped.mandataris} lmb:hasPublicationStatus ?status.
-        }
-        OPTIONAL {
-          ${escaped.mandataris} dct:modified ?modified.
-        }
-      }
-      BIND(NOW() as ?now)
-    }
-  `;
-
-  try {
-    await updateSudo(updateStatusQuery);
-    console.log(
-      `|> Updated status to ${status} for mandataris: ${mandataris}.`,
-    );
-  } catch (error) {
-    console.log(
-      `|> Could not update mandataris: ${mandataris} status to ${status}`,
-    );
-  }
-}
-
 export async function bulkSetPublicationStatusNietBekrachtigd(
   mandatarissen: string[],
 ): Promise<void> {
