+++ data-access/mandatees-decisions.ts
@@ -44,41 +44,3 @@ export async function getGraphsWhereInstanceExists(instanceUri) {
 
   return getSparqlResults(result).map((b) => b.graph.value);
 }
-
-export async function addBesluitToMandataris(
-  mandatarisUri: string,
-  besluitUri: string,
-  link: string,
-  graphs: Array<string>,
-) {
-  if (graphs.length === 0) {
-    console.log(
-      `|> No graphs found to insert mandataris: ${mandatarisUri} with besluit: ${besluitUri}`,
-    );
-    return;
-  }
-
-  const escaped = {
-    mandataris: sparqlEscapeUri(mandatarisUri),
-    link: sparqlEscapeUri(link),
-    besluit: sparqlEscapeUri(besluitUri),
-  };
-  const insertString = `
-    PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
-    PREFIX mandaat: <http://data.vlaanderen.be/ns/mandaat#>
-
-    INSERT {
-      GRAPH ?g {
-        ${escaped.mandataris} ${escaped.link} ${escaped.besluit} .
-      }
-    } WHERE {
-      VALUES ?g { ${graphs.map((g) => sparqlEscapeUri(g))} }
-      GRAPH ?g {
-        ${escaped.mandataris} a mandaat:Mandataris .
-      }
-      ?g ext:ownedBy ?bestuurseenheid . 
-    }
-  `;
-
-  await updateSudo(insertString, graphs);
-}
